requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
sqlalchemy==2.0.25
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
